@@ -5,10 +5,10 @@ namespace Model
     {
 
         public string DiscoveryUrl { get; private set; } = string.Empty;
-        public List<Error> Errors { get; private set; } = new List<Error>();
+        public List<Error> Errors { get; private set; } = [];
         public ICollection<Issue> Issues { get; private set; } = new List<Issue>();
         public ICollection<ISecurityTestSession> securityTestSessions { get; private set; } = new List<ISecurityTestSession>();
-        public EndpointDescriptionCollection EndpointDescriptions { get; private set; } = new EndpointDescriptionCollection();
+        public EndpointDescriptionCollection EndpointDescriptions { get; private set; } = [];
 
 
         public Server(string DiscoveryUrl, EndpointDescriptionCollection edc)
