@@ -15,7 +15,7 @@ public class CommonCredentialsPluginTest
     private readonly Mock<ISecurityTestSession> _mockSession;
     private readonly Mock<ISecurityTestSession> _mockSessionSuccess;
     private readonly string _discoveryUrl = "opc.tcp://localhost:4840";
-    private readonly EndpointDescriptionCollection _endpointDescriptions = new();
+    private readonly EndpointDescriptionCollection _endpointDescriptions = [];
 
     public CommonCredentialsPluginTest()
     {
