@@ -12,36 +12,38 @@ using Xunit;
 namespace Tests;
 public class DiscoveryControllerTest
 {
-    private readonly ApplicationDescriptionCollection _validApplicationDescriptionCollection = new() { new ApplicationDescription()
+    private readonly ApplicationDescriptionCollection _validApplicationDescriptionCollection = [ new ApplicationDescription()
     {
         ApplicationName = "test",
         ApplicationType = ApplicationType.Server,
         ApplicationUri = "asd",
         ProductUri = "asd",
-        DiscoveryUrls = new StringCollection()
-            {
+        DiscoveryUrls =
+            [
                 "opc.tcp://example.com:53530"
-            }
-    } };
-    private readonly EndpointDescriptionCollection _validEndpointDescriptionCollection = new()
-    {
+            ]
+    } ];
+    private readonly EndpointDescriptionCollection _validEndpointDescriptionCollection =
+    [
         new EndpointDescription()
         {
             EndpointUrl = "asd"
         }
-    };
+    ];
 
-    private readonly ILogger _logger;
-    private readonly List<Uri> _discoveryUris = new()
-    {
+    private readonly List<Uri> _discoveryUris =
+    [
         new Uri("opc.tcp://asd.com"),
-    };
+    ];
     private readonly Mock<IDiscoveryUtil> _mockDiscoveryUtil;
+    private readonly Mock<ITaskUtil> _mockTaskUtil;
+    private readonly Mock<ILogger<DiscoveryController>> _loggerMock;
 
     public DiscoveryControllerTest()
     {
-        _logger = LoggerFactory.Create(builder => { }).CreateLogger<DiscoveryControllerTest>();
+        _loggerMock = new Mock<ILogger<DiscoveryController>>();
         _mockDiscoveryUtil = new Mock<IDiscoveryUtil>();
+        _mockTaskUtil = new Mock<ITaskUtil>();
     }
 
 
@@ -53,7 +55,7 @@ public class DiscoveryControllerTest
         _mockDiscoveryUtil.Setup(util => util.DiscoverApplications(It.IsAny<Uri>())).Returns(_validApplicationDescriptionCollection);
         _mockDiscoveryUtil.Setup(util => util.DiscoverEndpoints(It.IsAny<Uri>())).Returns(_validEndpointDescriptionCollection);
 
-        DiscoveryController controller = new(_logger, _mockDiscoveryUtil.Object);
+        DiscoveryController controller = new(_loggerMock.Object, _mockDiscoveryUtil.Object, _mockTaskUtil.Object);
 
         // act
         ICollection<Target> targets = controller.DiscoverTargets(_discoveryUris);
@@ -71,8 +73,8 @@ public class DiscoveryControllerTest
     {
         // arrange
         DiscoveryUtil discoveryUtil = new();
-        DiscoveryController controller = new(_logger, discoveryUtil);
-        List<Uri> discoveryUris = new();
+        DiscoveryController controller = new(_loggerMock.Object, discoveryUtil, _mockTaskUtil.Object);
+        List<Uri> discoveryUris = [];
 
         // act
         ICollection<Target> targets = controller.DiscoverTargets(discoveryUris);
@@ -85,10 +87,10 @@ public class DiscoveryControllerTest
     public void HttpsDiscoveryUriReturnsEmptyTargets()
     {
         // arrange
-        DiscoveryController controller = new(_logger, _mockDiscoveryUtil.Object);
-        List<Uri> discoveryUris = new() {
+        DiscoveryController controller = new(_loggerMock.Object, _mockDiscoveryUtil.Object, _mockTaskUtil.Object);
+        List<Uri> discoveryUris = [
             new Uri("https://test.com")
-        };
+        ];
 
         // act
         ICollection<Target> targets = controller.DiscoverTargets(discoveryUris);
@@ -105,7 +107,7 @@ public class DiscoveryControllerTest
         _mockDiscoveryUtil.Setup(util => util.DiscoverApplications(It.IsAny<Uri>())).Returns(_validApplicationDescriptionCollection);
         _mockDiscoveryUtil.Setup(util => util.DiscoverEndpoints(It.IsAny<Uri>())).Returns(_validEndpointDescriptionCollection);
 
-        DiscoveryController controller = new(_logger, _mockDiscoveryUtil.Object);
+        DiscoveryController controller = new(_loggerMock.Object, _mockDiscoveryUtil.Object, _mockTaskUtil.Object);
 
         // act
         ICollection<Target> targets = controller.DiscoverTargets(_discoveryUris);
@@ -122,7 +124,7 @@ public class DiscoveryControllerTest
         _mockDiscoveryUtil.Setup(util => util.DiscoverApplications(It.IsAny<Uri>())).Returns(_validApplicationDescriptionCollection);
         _mockDiscoveryUtil.Setup(util => util.DiscoverEndpoints(It.IsAny<Uri>())).Returns(_validEndpointDescriptionCollection);
 
-        DiscoveryController controller = new(_logger, _mockDiscoveryUtil.Object);
+        DiscoveryController controller = new(_loggerMock.Object, _mockDiscoveryUtil.Object, _mockTaskUtil.Object);
 
         // act
         ICollection<Target> targets = controller.DiscoverTargets(_discoveryUris);
@@ -135,11 +137,11 @@ public class DiscoveryControllerTest
     public void ResolvesToZeroAddresses()
     {
         // arrange
-        _mockDiscoveryUtil.Setup(util => util.ResolveIPv4Addresses(It.IsAny<string>())).Returns(Array.Empty<IPAddress>());
+        _mockDiscoveryUtil.Setup(util => util.ResolveIPv4Addresses(It.IsAny<string>())).Returns([]);
         _mockDiscoveryUtil.Setup(util => util.DiscoverApplications(It.IsAny<Uri>())).Returns(_validApplicationDescriptionCollection);
         _mockDiscoveryUtil.Setup(util => util.DiscoverEndpoints(It.IsAny<Uri>())).Returns(_validEndpointDescriptionCollection);
 
-        DiscoveryController controller = new(_logger, _mockDiscoveryUtil.Object);
+        DiscoveryController controller = new(_loggerMock.Object, _mockDiscoveryUtil.Object, _mockTaskUtil.Object);
 
         // act
         ICollection<Target> targets = controller.DiscoverTargets(_discoveryUris);
@@ -156,7 +158,7 @@ public class DiscoveryControllerTest
         _mockDiscoveryUtil.Setup(util => util.DiscoverApplications(It.IsAny<Uri>())).Throws<ServiceResultException>();
         _mockDiscoveryUtil.Setup(util => util.DiscoverEndpoints(It.IsAny<Uri>())).Returns(_validEndpointDescriptionCollection);
 
-        DiscoveryController controller = new(_logger, _mockDiscoveryUtil.Object);
+        DiscoveryController controller = new(_loggerMock.Object, _mockDiscoveryUtil.Object, _mockTaskUtil.Object);
 
         // act
         ICollection<Target> targets = controller.DiscoverTargets(_discoveryUris);
@@ -173,7 +175,7 @@ public class DiscoveryControllerTest
         _mockDiscoveryUtil.Setup(util => util.DiscoverApplications(It.IsAny<Uri>())).Throws<Exception>();
         _mockDiscoveryUtil.Setup(util => util.DiscoverEndpoints(It.IsAny<Uri>())).Returns(_validEndpointDescriptionCollection);
 
-        DiscoveryController controller = new(_logger, _mockDiscoveryUtil.Object);
+        DiscoveryController controller = new(_loggerMock.Object, _mockDiscoveryUtil.Object, _mockTaskUtil.Object);
 
         // act & assert
         Assert.Throws<Exception>(() => controller.DiscoverTargets(_discoveryUris));
@@ -187,7 +189,7 @@ public class DiscoveryControllerTest
         _mockDiscoveryUtil.Setup(util => util.DiscoverEndpoints(It.IsAny<Uri>())).Throws<ServiceResultException>();
         _mockDiscoveryUtil.Setup(util => util.DiscoverApplications(It.IsAny<Uri>())).Returns(_validApplicationDescriptionCollection);
 
-        DiscoveryController controller = new(_logger, _mockDiscoveryUtil.Object);
+        DiscoveryController controller = new(_loggerMock.Object, _mockDiscoveryUtil.Object, _mockTaskUtil.Object);
 
         // act
         ICollection<Target> targets = controller.DiscoverTargets(_discoveryUris);
@@ -205,7 +207,7 @@ public class DiscoveryControllerTest
         _mockDiscoveryUtil.Setup(util => util.DiscoverEndpoints(It.IsAny<Uri>())).Throws<Exception>();
         _mockDiscoveryUtil.Setup(util => util.DiscoverApplications(It.IsAny<Uri>())).Returns(_validApplicationDescriptionCollection);
 
-        DiscoveryController controller = new(_logger, _mockDiscoveryUtil.Object);
+        DiscoveryController controller = new(_loggerMock.Object, _mockDiscoveryUtil.Object, _mockTaskUtil.Object);
 
         // act & assert
         Assert.Throws<Exception>(() => controller.DiscoverTargets(_discoveryUris));
