@@ -9,7 +9,8 @@ using Xunit;
 namespace Tests;
 public class ReportControllerTest
 {
-    private readonly Mock<ILogger> _loggerMock;
+    private readonly MemoryStream _outputStream = new();
+    private readonly Mock<ILogger<IReportController>> _loggerMock;
     private readonly EndpointDescription _endpointDescription = new()
     {
         UserIdentityTokens = new UserTokenPolicyCollection(new List<UserTokenPolicy> { new(UserTokenType.UserName) }),
@@ -28,9 +29,9 @@ public class ReportControllerTest
 
     public ReportControllerTest()
     {
-        _loggerMock = new Mock<ILogger>();
+        _loggerMock = new Mock<ILogger<IReportController>>();
         _reporterMock = new Mock<IReporter>();
-        _reporterMock.Setup(r => r.PrintXHTMLReport(It.IsAny<Report>())).Verifiable();
+        _reporterMock.Setup(r => r.PrintXHTMLReport(It.IsAny<Report>(), _outputStream)).Verifiable();
     }
 
     // add test to check if targets are sorted by server issue severity
@@ -40,13 +41,14 @@ public class ReportControllerTest
         // Arrange
 
         // Act
-        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object, new List<Target>(), DateTime.Now, DateTime.Now, string.Empty, string.Empty);
-        reportController.WriteReport();
+        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object);
+        Report report = reportController.GenerateReport(new List<Target>(), DateTime.Now, DateTime.Now, "", "");
+        reportController.WriteReport(report, _outputStream);
 
         // Assert
-        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>()), Times.Once);
-        Assert.NotNull(reportController.report);
-        Assert.Empty(reportController.report.Targets);
+        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>(), _outputStream), Times.Once);
+        Assert.NotNull(report);
+        Assert.Empty(report.Targets);
     }
 
     [Fact]
@@ -55,14 +57,15 @@ public class ReportControllerTest
         // Arrange
 
         // Act
-        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object, new List<Target>() { new(_applicationDescription) }, DateTime.Now, DateTime.Now, string.Empty, string.Empty);
-        reportController.WriteReport();
+        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object);
+        Report report = reportController.GenerateReport(new List<Target> { new(_applicationDescription) }, DateTime.Now, DateTime.Now, "", "");
+        reportController.WriteReport(report, _outputStream);
 
         // Assert
-        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>()), Times.Once);
-        Assert.NotNull(reportController.report);
-        Assert.NotEmpty(reportController.report.Targets);
-        Assert.True(reportController.report.Targets.Count == 1);
+        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>(), _outputStream), Times.Once);
+        Assert.NotNull(report);
+        Assert.NotEmpty(report.Targets);
+        Assert.True(report.Targets.Count == 1);
     }
 
     // add test to check if targets are sorted by server issue severity
@@ -72,28 +75,29 @@ public class ReportControllerTest
         // Arrange
 
         var target1 = new Target(_applicationDescription);
-        target1.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
+        target1.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
         target1.Servers.First().Issues.Add(new Issue(1, "description", 0.1));
 
         var target2 = new Target(_applicationDescription);
-        target2.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
+        target2.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
         target2.Servers.First().Issues.Add(new Issue(1, "description", 0.2));
 
         var target3 = new Target(_applicationDescription);
-        target3.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
+        target3.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
         target3.Servers.First().Issues.Add(new Issue(1, "description", 0.3));
 
         // Act
-        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object, new List<Target>() { target1, target2, target3 }, DateTime.Now, DateTime.Now, string.Empty, string.Empty);
-        reportController.WriteReport();
+        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object);
+        Report report = reportController.GenerateReport(new List<Target> { target1, target2, target3 }, DateTime.Now, DateTime.Now, "", "");
+        reportController.WriteReport(report, _outputStream);
 
         // Assert
-        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>()), Times.Once);
-        Assert.NotNull(reportController.report);
-        Assert.NotEmpty(reportController.report.Targets);
-        Assert.True(reportController.report.Targets.Count == 3);
-        Assert.True(reportController.report.Targets.First().Servers.First().Issues.First().Severity == 0.3);
-        Assert.True(reportController.report.Targets.Last().Servers.First().Issues.First().Severity == 0.1);
+        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>(), _outputStream), Times.Once);
+        Assert.NotNull(report);
+        Assert.NotEmpty(report.Targets);
+        Assert.True(report.Targets.Count == 3);
+        Assert.True(report.Targets.First().Servers.First().Issues.First().Severity == 0.3);
+        Assert.True(report.Targets.Last().Servers.First().Issues.First().Severity == 0.1);
     }
 
     // test that report is generated correctly when not all targets have servers
@@ -103,25 +107,26 @@ public class ReportControllerTest
         // Arrange
 
         var target1 = new Target(_applicationDescription);
-        target1.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
+        target1.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
         target1.Servers.First().Issues.Add(new Issue(1, "description", 0.1));
 
         var target2 = new Target(_applicationDescription);
 
         var target3 = new Target(_applicationDescription);
-        target3.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
+        target3.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
         target3.Servers.First().Issues.Add(new Issue(1, "description", 0.3));
 
         // Act
-        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object, new List<Target>() { target1, target2, target3 }, DateTime.Now, DateTime.Now, string.Empty, string.Empty);
-        reportController.WriteReport();
+        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object);
+        Report report = reportController.GenerateReport(new List<Target> { target1, target2, target3 }, DateTime.Now, DateTime.Now, "", "");
+        reportController.WriteReport(report, _outputStream);
 
         // Assert
-        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>()), Times.Once);
-        Assert.NotNull(reportController.report);
-        Assert.NotEmpty(reportController.report.Targets);
-        Assert.True(reportController.report.Targets.Count == 3);
-        Assert.True(reportController.report.Targets.First().Servers.First().Issues.First().Severity == 0.3);
+        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>(), _outputStream), Times.Once);
+        Assert.NotNull(report);
+        Assert.NotEmpty(report.Targets);
+        Assert.True(report.Targets.Count == 3);
+        Assert.True(report.Targets.First().Servers.First().Issues.First().Severity == 0.3);
     }
 
     // test that report is generated correctly when not all servers have endpoints
@@ -131,26 +136,27 @@ public class ReportControllerTest
         // Arrange
 
         var target1 = new Target(_applicationDescription);
-        target1.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
+        target1.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
         target1.Servers.First().Issues.Add(new Issue(1, "description", 0.1));
 
         var target2 = new Target(_applicationDescription);
-        target2.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { }));
+        target2.AddServer(new Server("opc.tcp://discoveryuri", []));
 
         var target3 = new Target(_applicationDescription);
-        target3.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
+        target3.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
         target3.Servers.First().Issues.Add(new Issue(1, "description", 0.3));
 
         // Act
-        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object, new List<Target>() { target1, target2, target3 }, DateTime.Now, DateTime.Now, string.Empty, string.Empty);
-        reportController.WriteReport();
+        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object);
+        Report report = reportController.GenerateReport(new List<Target> { target1, target2, target3 }, DateTime.Now, DateTime.Now, "", "");
+        reportController.WriteReport(report, _outputStream);
 
         // Assert
-        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>()), Times.Once);
-        Assert.NotNull(reportController.report);
-        Assert.NotEmpty(reportController.report.Targets);
-        Assert.True(reportController.report.Targets.Count == 3);
-        Assert.True(reportController.report.Targets.First().Servers.First().Issues.First().Severity == 0.3);
+        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>(), _outputStream), Times.Once);
+        Assert.NotNull(report);
+        Assert.NotEmpty(report.Targets);
+        Assert.True(report.Targets.Count == 3);
+        Assert.True(report.Targets.First().Servers.First().Issues.First().Severity == 0.3);
     }
 
     // test that report is generated correctly when not all endpoints have issues
@@ -160,30 +166,31 @@ public class ReportControllerTest
         // Arrange
 
         var target1 = new Target(_applicationDescription);
-        target1.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
+        target1.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
         target1.Servers.First().Issues.Add(new Issue(1, "description", 0.1));
         target1.Servers.First().Issues.Add(new Issue(2, "description", 0.1));
         target1.Servers.First().Issues.Add(new Issue(3, "description", 0.1));
 
         var target2 = new Target(_applicationDescription);
-        target2.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
-        target2.AddServer(new Server("opc.tcp://discoveryuri2", new EndpointDescriptionCollection() { _endpointDescription }));
-        target2.AddServer(new Server("opc.tcp://discoveryuri3", new EndpointDescriptionCollection() { _endpointDescription }));
+        target2.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
+        target2.AddServer(new Server("opc.tcp://discoveryuri2", [_endpointDescription]));
+        target2.AddServer(new Server("opc.tcp://discoveryuri3", [_endpointDescription]));
 
         var target3 = new Target(_applicationDescription);
-        target3.AddServer(new Server("opc.tcp://discoveryuri", new EndpointDescriptionCollection() { _endpointDescription }));
+        target3.AddServer(new Server("opc.tcp://discoveryuri", [_endpointDescription]));
         target3.Servers.First().Issues.Add(new Issue(1, "description", 0.3));
 
         // Act
-        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object, new List<Target>() { target1, target2, target3 }, DateTime.Now, DateTime.Now, string.Empty, string.Empty);
-        reportController.WriteReport();
+        ReportController reportController = new(_loggerMock.Object, _reporterMock.Object);
+        Report report = reportController.GenerateReport(new List<Target> { target1, target2, target3 }, DateTime.Now, DateTime.Now, "", "");
+        reportController.WriteReport(report, _outputStream);
 
         // Assert
-        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>()), Times.Once);
-        Assert.NotNull(reportController.report);
-        Assert.NotEmpty(reportController.report.Targets);
-        Assert.True(reportController.report.Targets.Count == 3);
-        Assert.True(reportController.report.Targets.First().Servers.First().Issues.First().Severity == 0.3);
+        _reporterMock.Verify(r => r.PrintXHTMLReport(It.IsAny<Report>(), _outputStream), Times.Once);
+        Assert.NotNull(report);
+        Assert.NotEmpty(report.Targets);
+        Assert.True(report.Targets.Count == 3);
+        Assert.True(report.Targets.First().Servers.First().Issues.First().Severity == 0.3);
     }
 
 }
