@@ -1,6 +1,7 @@
 using Controller;
 using Microsoft.Extensions.Logging;
 using Model;
+using Moq;
 using View;
 using Xunit;
 
@@ -10,19 +11,22 @@ public class ReportController_WriteReportShould
     [Fact]
     public void WriteReport_AddsReportRunStatus()
     {
+        Mock<ILogger<IReportController>> _loggerMock = new();
+
         ILoggerFactory loggerFactory = LoggerFactory.Create(builder => { });
         ILogger logger = loggerFactory.CreateLogger<ReportController_WriteReportShould>();
         StreamWriter sw = new(new MemoryStream())
         {
             AutoFlush = true
         };
-        Reporter reporter = new(sw.BaseStream);
+        Reporter reporter = new();
         string runStatus = "hello";
         string commandLine = "commandline";
-        ReportController reportController = new(logger, reporter, new List<Target>(), DateTime.Now, DateTime.Now, commandLine, runStatus);
-        reportController.WriteReport();
+        ReportController reportController = new(_loggerMock.Object, reporter);
+        Report report = reportController.GenerateReport(new List<Target>(), DateTime.Now, DateTime.Now, commandLine, runStatus);
+        reportController.WriteReport(report, new MemoryStream());
 
-        Assert.True(reportController.report!.RunStatus == runStatus);
-        Assert.True(reportController.report!.Command == commandLine);
+        Assert.True(report.RunStatus == runStatus);
+        Assert.True(report.Command == commandLine);
     }
 }
