@@ -11,7 +11,7 @@ public class SecurityPolicyBasic128Rsa15PluginTest
     private readonly ILogger _logger;
     private readonly SecurityPolicyBasic128Rsa15Plugin _plugin;
     private readonly string _discoveryUrl = "opc.tcp://localhost:4840";
-    private readonly EndpointDescriptionCollection _endpointDescriptions = new();
+    private readonly EndpointDescriptionCollection _endpointDescriptions = [];
 
     public SecurityPolicyBasic128Rsa15PluginTest()
     {
