@@ -11,7 +11,7 @@ public class SecurityPolicyBasic256PluginTest
     private readonly ILogger _logger;
     private readonly SecurityPolicyBasic256Plugin _plugin;
     private readonly string _discoveryUrl = "opc.tcp://localhost:4840";
-    private readonly EndpointDescriptionCollection _endpointDescriptions = new();
+    private readonly EndpointDescriptionCollection _endpointDescriptions = [];
     public SecurityPolicyBasic256PluginTest()
     {
         _logger = LoggerFactory.Create(builder => { }).CreateLogger<SecurityPolicyBasic256PluginTest>();
