@@ -11,7 +11,7 @@ public class SecurityPolicyNonePluginTest
     private readonly ILogger _logger;
     private readonly SecurityPolicyNonePlugin _plugin;
     private readonly string _discoveryUrl = "opc.tcp://localhost:4840";
-    private readonly EndpointDescriptionCollection _endpointDescriptions = new();
+    private readonly EndpointDescriptionCollection _endpointDescriptions = [];
 
     public SecurityPolicyNonePluginTest()
     {
