@@ -1,3 +1,4 @@
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Model;
 using Plugin;
@@ -6,52 +7,49 @@ using Util;
 namespace Controller
 {
 
-    public class SecurityTestController
+    public interface ISecurityTestController
     {
+        ICollection<Target> TestTargetSecurity(ICollection<Target> opcTargets, AuthenticationData authenticationData);
+    }
 
-        readonly ILogger _logger;
-        readonly ICollection<IPlugin> _securityTestPlugins;
-        readonly CancellationToken? _token;
-
-        public SecurityTestController(ILogger logger, ICollection<IPlugin> SecurityTestPlugins, CancellationToken? token = null)
-        {
-            _logger = logger;
-            _securityTestPlugins = SecurityTestPlugins;
-            _token = token;
-        }
+    public class SecurityTestController(ILogger<ISecurityTestController> logger, ITaskUtil taskUtil, IPluginRepository pluginRepository) : ISecurityTestController
+    {
+        private ICollection<IPlugin> _securityTestPlugins = [];
 
 
         // Run all security tests and return result-populated opcTarget
-        public ICollection<Target> TestTargetSecurity(ICollection<Target> opcTargets)
+        public ICollection<Target> TestTargetSecurity(ICollection<Target> opcTargets, AuthenticationData authenticationData)
         {
-            _logger.LogTrace("{Message}", $"Loaded {_securityTestPlugins.Count} security test plugins");
-            _logger.LogTrace("{Message}", $"Plugins: {string.Join(", ", _securityTestPlugins.Select(p => (int)p.pluginId))}");
+            _securityTestPlugins = pluginRepository.GetAll(authenticationData);
 
-            _logger.LogDebug("{Message}", $"Starting security tests of {opcTargets.Count} targets");
+            logger.LogTrace("{Message}", $"Loaded {_securityTestPlugins.Count} security test plugins");
+            logger.LogTrace("{Message}", $"Plugins: {string.Join(", ", _securityTestPlugins.Select(p => (int)p.pluginId))}");
+
+            logger.LogDebug("{Message}", $"Starting security tests of {opcTargets.Count} targets");
 
             Parallel.ForEach(opcTargets, new ParallelOptions { MaxDegreeOfParallelism = 10 }, target =>
             {
-                _logger.LogDebug("{Message}", $"Testing {target.ApplicationName} ({target.ProductUri})");
+                logger.LogDebug("{Message}", $"Testing {target.ApplicationName} ({target.ProductUri})");
 
                 try
                 {
                     foreach (Server server in target.Servers)
                     {
-                        _logger.LogTrace("{Message}", $"Testing endpoint {server.DiscoveryUrl} of {target.ApplicationName}");
+                        logger.LogTrace("{Message}", $"Testing endpoint {server.DiscoveryUrl} of {target.ApplicationName}");
                         TestEndpointSecurity(server);
 
-                        if (!server.securityTestSessions.Any())
+                        if (server.securityTestSessions.Count == 0)
                         {
-                            _logger.LogWarning("{Message}", $"Cannot authenticate to {target.ApplicationName}. Skipping post-authentication tests");
+                            logger.LogWarning("{Message}", $"Cannot authenticate to {target.ApplicationName}. Skipping post-authentication tests");
                         }
                     }
                 }
                 catch (Exception e)
                 {
                     string msg = $"Unknown exception scanning {target.ApplicationName}: {e}";
-                    _logger.LogError("{Message}", msg);
+                    logger.LogError("{Message}", msg);
 
-                    if (target.Servers.Any())
+                    if (target.Servers.Count != 0)
                     {
                         target.Servers.First().AddError(new Error(msg));
                     }
@@ -74,64 +72,61 @@ namespace Controller
                 server.AddSecurityTestSession(session);
             }
 
-            if (server.securityTestSessions.Any())
+            if (server.securityTestSessions.Count != 0)
             {
                 TestEndpointPostAuth(server);
             }
 
-            _logger.LogTrace("{Message}", $"Closing sessions");
+            logger.LogTrace("{Message}", $"Closing sessions");
             foreach (ISecurityTestSession securityTestSession in server.securityTestSessions)
             {
                 securityTestSession.Dispose();
             }
         }
 
-        private ICollection<ISecurityTestSession> TestEndpointPreauth(Server server)
+        private List<ISecurityTestSession> TestEndpointPreauth(Server server)
         {
-            List<ISecurityTestSession> securityTestSessions = new();
+            List<ISecurityTestSession> securityTestSessions = [];
 
-            _logger.LogTrace("{Message}", $"Starting pre-authentication tests");
+            logger.LogTrace("{Message}", $"Starting pre-authentication tests");
             foreach (IPreAuthPlugin preauthPlugin in GetPluginsByType(Plugintype.PreAuthPlugin).Cast<IPreAuthPlugin>())
             {
-                TaskUtil.CheckForCancellation(_token);
+                taskUtil.CheckForCancellation();
                 (Issue? preauthissue, ICollection<ISecurityTestSession> preauthsessions) = preauthPlugin.Run(server.DiscoveryUrl, server.EndpointDescriptions);
 
                 securityTestSessions.AddRange(preauthsessions);
                 if (preauthissue != null) server.AddIssue(preauthissue);
             }
-            _logger.LogTrace("{Message}", $"Finished pre-authentication tests");
+            logger.LogTrace("{Message}", $"Finished pre-authentication tests");
 
             return securityTestSessions;
         }
 
         private void TestEndpointPostAuth(Server server)
         {
-            _logger.LogTrace("{Message}", $"Starting post-authentication tests");
+            logger.LogTrace("{Message}", $"Starting post-authentication tests");
 
             TestEndpointSessionCredentials(server);
 
             foreach (IPostAuthPlugin postAuthPlugin in GetPluginsByType(Plugintype.PostAuthPlugin).Cast<IPostAuthPlugin>())
             {
-                TaskUtil.CheckForCancellation(_token);
+                taskUtil.CheckForCancellation();
                 Issue? postauthIssue = postAuthPlugin.Run(server.securityTestSessions.First().Session);
                 if (postauthIssue != null) server.AddIssue(postauthIssue);
             }
-            _logger.LogTrace("{Message}", $"Finished post-authentication tests");
+            logger.LogTrace("{Message}", $"Finished post-authentication tests");
         }
 
         private void TestEndpointSessionCredentials(Server server)
         {
             foreach (ISessionCredentialPlugin sessionCredentialPlugin in GetPluginsByType(Plugintype.SessionCredentialPlugin).Cast<ISessionCredentialPlugin>())
             {
-                TaskUtil.CheckForCancellation(_token);
+                taskUtil.CheckForCancellation();
                 Issue? postauthIssue = sessionCredentialPlugin.Run(server.securityTestSessions);
                 if (postauthIssue != null) server.AddIssue(postauthIssue);
             }
         }
 
-        private ICollection<IPlugin> GetPluginsByType(Plugintype plugintype)
-        {
-            return _securityTestPlugins.Where(p => p.Type == plugintype).ToList();
-        }
+        private List<IPlugin> GetPluginsByType(Plugintype plugintype) => _securityTestPlugins.Where(p => p.Type == plugintype).ToList();
     }
 }
