@@ -4,21 +4,21 @@ namespace Model
 {
     public class AuthenticationData
     {
-        public List<CertificateIdentifier> applicationCertificates = new();
-        public List<CertificateIdentifier> userCertificates = new();
+        public List<CertificateIdentifier> applicationCertificates = [];
+        public List<CertificateIdentifier> userCertificates = [];
 
-        public List<(string, string)> loginCredentials = new();
-        public List<(string, string)> bruteForceCredentials = new();
+        public List<(string, string)> loginCredentials = [];
+        public List<(string, string)> bruteForceCredentials = [];
 
         public AuthenticationData()
         {
         }
         public AuthenticationData(ICollection<CertificateIdentifier> applicationCertificates, ICollection<CertificateIdentifier> userCertificates, ICollection<(string, string)> loginCredentials, ICollection<(string, string)> bruteForceCredentials)
         {
-            this.applicationCertificates = applicationCertificates.ToList();
-            this.userCertificates = userCertificates.ToList();
-            this.loginCredentials = loginCredentials.ToList();
-            this.bruteForceCredentials = bruteForceCredentials.ToList();
+            this.applicationCertificates = [.. applicationCertificates];
+            this.userCertificates = [.. userCertificates];
+            this.loginCredentials = [.. loginCredentials];
+            this.bruteForceCredentials = [.. bruteForceCredentials];
         }
 
         public void AddLoginCredential(string username, string password)
