@@ -2,18 +2,10 @@ using Opc.Ua;
 
 namespace Model
 {
-    public class SessionCredential
+    public class SessionCredential(UserIdentity userIdentity, CertificateIdentifier appCertificateIdentifier, bool selfSigned = false)
     {
-        public UserIdentity identity { get; private set; }
-        public CertificateIdentifier applicationCertificate { get; private set; }
-        public bool selfSignedAppCert { get; private set; }
-
-
-        public SessionCredential(UserIdentity userIdentity, CertificateIdentifier appCertificateIdentifier, bool selfSigned = false)
-        {
-            identity = userIdentity;
-            applicationCertificate = appCertificateIdentifier;
-            selfSignedAppCert = selfSigned;
-        }
+        public UserIdentity identity { get; private set; } = userIdentity;
+        public CertificateIdentifier applicationCertificate { get; private set; } = appCertificateIdentifier;
+        public bool selfSignedAppCert { get; private set; } = selfSigned;
     }
 }
