@@ -2,32 +2,19 @@ using Opc.Ua;
 
 namespace Model
 {
-    public class Target
+    public class Target(ApplicationDescription ad)
     {
 
-        private readonly ApplicationDescription _applicationDescription = new();
-        public List<Server> Servers { get; private set; } = new();
+        private readonly ApplicationDescription _applicationDescription = ad;
+        public List<Server> Servers { get; private set; } = [];
 
-        public ApplicationType Type { get; private set; }
-        public string ApplicationName { get; private set; }
-        public string ApplicationUri { get; private set; }
-        public string ProductUri { get; private set; }
+        public ApplicationType Type { get; private set; } = ad.ApplicationType;
+        public string ApplicationName { get; private set; } = ad.ApplicationName.ToString();
+        public string ApplicationUri { get; private set; } = ad.ApplicationUri;
+        public string ProductUri { get; private set; } = ad.ProductUri;
 
         public int IssuesCount => Servers.Sum(s => s.Issues.Count);
         public int ErrorsCount => Servers.Sum(s => s.Errors.Count);
-
-
-        public Target(ApplicationDescription ad)
-        {
-            _applicationDescription = ad;
-
-            Type = ad.ApplicationType;
-            ApplicationName = ad.ApplicationName.ToString();
-            ApplicationUri = ad.ApplicationUri;
-            ProductUri = ad.ProductUri;
-
-            Servers = new List<Server>();
-        }
 
         public void AddServer(Server server)
         {
@@ -37,7 +24,7 @@ namespace Model
         public void SortServersByIssueSeverity()
         {
             // sort servers within target by issue severity even if issues is empty
-            Servers = Servers.OrderByDescending(s => s.Issues.Any() ? s.Issues.Max(i => i.Severity) : 0).ToList();
+            Servers = [.. Servers.OrderByDescending(s => s.Issues.Count != 0 ? s.Issues.Max(i => i.Severity) : 0)];
         }
     }
 }
